+++ src/maestral/sync.py
@@ -103,17 +103,19 @@ from .utils.integration import (
 )
 from .utils.path import (
     generate_cc_name,
-    cased_path_candidates,
-    is_fs_case_sensitive,
     move,
     delete,
     is_child,
     is_equal_or_child,
     content_hash,
     walk,
+    normalize,
+    normalize_case,
+    normalize_unicode,
+    equivalent_path_candidates,
 )
 from .utils.orm import Database, Manager
-from .utils.appdirs import get_data_path, get_home_dir
+from .utils.appdirs import get_data_path
 
 
 __all__ = [
@@ -537,7 +539,6 @@ class SyncEngine:
         self._mignore_path = osp.join(self._dropbox_path, MIGNORE_FILE)
         self._file_cache_path = osp.join(self._dropbox_path, FILE_CACHE)
 
-        self._is_case_sensitive = is_fs_case_sensitive(get_home_dir())
         self._excluded_items = self._conf.get("main", "excluded_items")
         self._max_cpu_percent = self._conf.get("sync", "max_cpu_percent") * CPU_COUNT
         self._local_cursor = self._state.get("sync", "lastsync")
@@ -566,7 +567,6 @@ class SyncEngine:
             self._dropbox_path = path
             self._mignore_path = osp.join(self._dropbox_path, MIGNORE_FILE)
             self._file_cache_path = osp.join(self._dropbox_path, FILE_CACHE)
-            self._is_case_sensitive = is_fs_case_sensitive(self._dropbox_path)
             self._conf.set("main", "path", path)
 
     @property
@@ -610,7 +610,7 @@ class SyncEngine:
         """
 
         # remove duplicate entries by creating set, strip trailing '/'
-        folder_set = {f.lower().rstrip("/") for f in folder_list}
+        folder_set = {normalize(f).rstrip("/") for f in folder_list}
 
         # remove all children of excluded folders
         clean_list = list(folder_set)
@@ -749,31 +749,31 @@ class SyncEngine:
         with self._database_access():
             return self._db_manager_index.count()
 
-    def get_local_rev(self, dbx_path: str) -> Optional[str]:
+    def get_local_rev(self, dbx_path_lower: str) -> Optional[str]:
         """
         Gets revision number of local file.
 
-        :param dbx_path: Dropbox path.
+        :param dbx_path_lower: Normalized lower case Dropbox path.
         :returns: Revision number as str or ``None`` if no local revision number has
             been saved.
         """
 
-        entry = self.get_index_entry(dbx_path)
+        entry = self.get_index_entry(dbx_path_lower)
 
         if entry:
             return entry.rev
         else:
             return None
 
-    def get_last_sync(self, dbx_path: str) -> float:
+    def get_last_sync(self, dbx_path_lower: str) -> float:
         """
         Returns the timestamp of last sync for an individual path.
 
-        :param dbx_path: Dropbox path.
+        :param dbx_path_lower: Normalized lower case Dropbox path.
         :returns: Time of last sync.
         """
 
-        entry = self.get_index_entry(dbx_path)
+        entry = self.get_index_entry(dbx_path_lower)
 
         if entry:
             last_sync = entry.last_sync or 0.0
@@ -782,16 +782,16 @@ class SyncEngine:
 
         return max(last_sync, self.local_cursor)
 
-    def get_index_entry(self, dbx_path: str) -> Optional[IndexEntry]:
+    def get_index_entry(self, dbx_path_lower: str) -> Optional[IndexEntry]:
         """
         Gets the index entry for the given Dropbox path.
 
-        :param dbx_path: Dropbox path.
+        :param dbx_path_lower: Normalized lower case Dropbox path.
         :returns: Index entry or ``None`` if no entry exists for the given path.
         """
 
         with self._database_access():
-            entry = self._db_manager_index.get(dbx_path.lower())
+            entry = self._db_manager_index.get(dbx_path_lower)
             return cast(Optional[IndexEntry], entry)
 
     def get_local_hash(self, local_path: str) -> Optional[str]:
@@ -890,7 +890,7 @@ class SyncEngine:
         if event.change_type is not ChangeType.Removed and not event.rev:
             raise ValueError("Rev required to update index")
 
-        dbx_path_lower = event.dbx_path.lower()
+        dbx_path_lower = event.dbx_path_lower
 
         with self._database_access():
 
@@ -899,7 +899,7 @@ class SyncEngine:
             if event.change_type is ChangeType.Removed:
                 self.remove_node_from_index(dbx_path_lower)
             elif event.change_type is ChangeType.Moved:
-                self.remove_node_from_index(event.dbx_path_from.lower())
+                self.remove_node_from_index(event.dbx_path_from_lower)
 
             # add or update entries for created or modified items
 
@@ -989,16 +989,16 @@ class SyncEngine:
 
                     self._db_manager_index.save(entry)
 
-    def remove_node_from_index(self, dbx_path: str) -> None:
+    def remove_node_from_index(self, dbx_path_lower: str) -> None:
         """
         Removes any local index entries for the given path and all its children.
 
-        :param dbx_path: Dropbox path.
+        :param dbx_path_lower: Normalized lower case Dropbox path.
         """
 
         with self._database_access():
 
-            dbx_path_lower = dbx_path.lower().rstrip("/")
+            dbx_path_lower = dbx_path_lower.rstrip("/")
 
             try:
                 self._db.execute(
@@ -1049,12 +1049,6 @@ class SyncEngine:
         self._mignore_rules = PathSpec.from_lines("gitwildmatch", spec.splitlines())
 
     # ==== helper functions ============================================================
-
-    @property
-    def is_case_sensitive(self) -> bool:
-        """Returns ``True`` if the local Dropbox folder is located on a partition with a
-        case-sensitive file system, ``False`` otherwise."""
-        return self._is_case_sensitive
 
     def ensure_dropbox_folder_present(self) -> None:
         """
@@ -1182,21 +1176,27 @@ class SyncEngine:
         :returns: Correctly cased Dropbox path.
         """
 
+        dbx_path_lower = normalize(dbx_path)
+
         client = client or self.client
 
         dirname, basename = osp.split(dbx_path)
+        dirname_lower = osp.dirname(dbx_path_lower)
 
-        dirname_cased = self._correct_case_helper(dirname, client)
+        dirname_cased = self._correct_case_helper(dirname, dirname_lower, client)
         path_cased = osp.join(dirname_cased, basename)
 
         # add our result to the cache
-        self._case_conversion_cache.put(dbx_path.lower(), path_cased)
+        self._case_conversion_cache.put(dbx_path_lower, path_cased)
 
         return path_cased
 
-    def _correct_case_helper(self, dbx_path: str, client: DropboxClient) -> str:
+    def _correct_case_helper(
+        self, dbx_path: str, dbx_path_lower: str, client: DropboxClient
+    ) -> str:
         """
         :param dbx_path: Uncased or randomly cased Dropbox path.
+        :param dbx_path_lower: Normalized fully lower cased Dropbox path.
         :param client: Client instance to use.
         :returns: Correctly cased Dropbox path.
         """
@@ -1206,7 +1206,6 @@ class SyncEngine:
             return dbx_path
 
         # check in our conversion cache
-        dbx_path_lower = dbx_path.lower()
         dbx_path_cased = self._case_conversion_cache.get(dbx_path_lower)
 
         if dbx_path_cased:
@@ -1220,7 +1219,7 @@ class SyncEngine:
             dbx_path_cased = entry.dbx_path_cased
         else:
             # fall back to querying from server
-            md = client.get_metadata(dbx_path_lower)
+            md = client.get_metadata(dbx_path)
             if md:
                 # recurse over parent directories
                 dbx_path_cased = self.correct_case(md.path_display, client)
@@ -1246,6 +1245,18 @@ class SyncEngine:
         if not is_equal_or_child(local_path, self.dropbox_path):
             raise ValueError(f'"{local_path}" is not in "{self.dropbox_path}"')
         return "/" + removeprefix(local_path, self.dropbox_path).lstrip("/")
+
+    def to_dbx_path_lower(self, local_path: str) -> str:
+        """
+        Converts a local path to a path relative to the Dropbox folder. The path will be
+        normalized as on Dropbox servers (lower case and some additional
+        normalisations).
+
+        :param local_path: Absolute path on local drive.
+        :returns: Relative path with respect to Dropbox folder.
+        :raises ValueError: When the path lies outside of the local Dropbox folder.
+        """
+        return normalize(self.to_dbx_path(local_path))
 
     def to_local_path_from_cased(self, dbx_path_cased: str) -> str:
         """
@@ -1301,12 +1312,12 @@ class SyncEngine:
             return
 
         dbx_path = cast(str, dbx_path)
-        dbx_path_lower = dbx_path.lower()
+        dbx_path_lower = normalize(dbx_path)
 
         if self.has_sync_errors():
             for error in self.sync_errors.copy():
                 if error.dbx_path and is_equal_or_child(
-                    error.dbx_path.lower(), dbx_path_lower
+                    normalize(error.dbx_path), dbx_path_lower
                 ):
                     try:
                         self.sync_errors.remove(error)
@@ -1323,7 +1334,7 @@ class SyncEngine:
         self.download_errors.clear()
 
     @staticmethod
-    def is_excluded(path) -> bool:
+    def is_excluded(path: str) -> bool:
         """
         Checks if a file is excluded from sync. Certain file names are always excluded
         from syncing, following the Dropbox support article:
@@ -1334,10 +1345,10 @@ class SyncEngine:
         temporary files as well as caches used by Dropbox or Maestral. `is_excluded`
         accepts both local and Dropbox paths.
 
-        :param path: Path of item. Can be both a local or Dropbox paths.
+        :param path: Can be either absolute or relative to the Dropbox folder. Does not
+            need to be normalized
         :returns: Whether the path is excluded from syncing.
         """
-        path = path.lower()
 
         # is root folder?
         if path in ("/", ""):
@@ -1349,9 +1360,7 @@ class SyncEngine:
             return True
 
         # in excluded dirs?
-        # TODO: check if this can be optimised
-        dirnames = dirname.split("/")
-        if any(excluded_dirname in dirnames for excluded_dirname in EXCLUDED_DIR_NAMES):
+        if dirname.split("/", 2)[1] in EXCLUDED_DIR_NAMES:
             return True
 
         if "~" in basename:  # is temporary file?
@@ -1366,16 +1375,15 @@ class SyncEngine:
 
         return False
 
-    def is_excluded_by_user(self, dbx_path: str) -> bool:
+    def is_excluded_by_user(self, dbx_path_lower: str) -> bool:
         """
         Check if file has been excluded through "selective sync" by the user.
 
-        :param dbx_path: Path relative to Dropbox folder.
+        :param dbx_path_lower: Normalised lower case Dropbox path.
         :returns: Whether the path is excluded from download syncing by the user.
         """
-        dbx_path = dbx_path.lower()
 
-        return any(is_equal_or_child(dbx_path, path) for path in self.excluded_items)
+        return any(is_equal_or_child(dbx_path_lower, p) for p in self.excluded_items)
 
     def is_mignore(self, event: SyncEvent) -> bool:
         """
@@ -1389,7 +1397,7 @@ class SyncEngine:
 
         return self._is_mignore_path(
             event.dbx_path, is_dir=event.is_directory
-        ) and not self.get_local_rev(event.dbx_path)
+        ) and not self.get_local_rev(event.dbx_path_lower)
 
     def _is_mignore_path(self, dbx_path: str, is_dir: bool = False) -> bool:
 
@@ -1496,9 +1504,9 @@ class SyncEngine:
 
             # save download errors to retry later
             if direction == SyncDirection.Down:
-                self.download_errors.add(err.dbx_path.lower())
+                self.download_errors.add(normalize(err.dbx_path))
             elif direction == SyncDirection.Up:
-                self.upload_errors.add(err.dbx_path.lower())
+                self.upload_errors.add(normalize(err.dbx_path))
 
     @contextmanager
     def _database_access(self, raise_error: bool = True) -> Iterator[None]:
@@ -1614,7 +1622,7 @@ class SyncEngine:
         for path, stat in walk(self.dropbox_path, self._scandir_with_ignore):
 
             is_dir = S_ISDIR(stat.st_mode)
-            dbx_path_lower = self.to_dbx_path(path).lower()
+            dbx_path_lower = self.to_dbx_path_lower(path)
             index_entry = self.get_index_entry(dbx_path_lower)
 
             if index_entry:
@@ -2048,8 +2056,8 @@ class SyncEngine:
         if (
             self.is_excluded(event.src_path)
             or self.is_excluded(event.dest_path)
-            or self.is_excluded_by_user(dbx_src_path)
-            or self.is_excluded_by_user(dbx_dest_path)
+            or self.is_excluded_by_user(normalize(dbx_src_path))
+            or self.is_excluded_by_user(normalize(dbx_dest_path))
         ):
             return True
 
@@ -2061,30 +2069,39 @@ class SyncEngine:
 
             return src_is_mignore or dest_is_mignore
 
-    def _handle_case_conflict(self, event: SyncEvent) -> bool:
+    def _handle_normalization_conflict(self, event: SyncEvent) -> bool:
         """
-        Checks for other items in the same directory with same name but a different
-        case. Renames items if necessary. Only needed for case sensitive file systems.
+        Checks for other items in the same directory with a different name but the same
+        normalization.
 
         :param event: SyncEvent for local created or moved event.
         :returns: Whether a case conflict was detected and handled.
         """
 
-        if not self.is_case_sensitive:
-            return False
-
         if not (event.is_added or event.is_moved):
             return False
 
         dirname, basename = osp.split(event.local_path)
+        equivalent_paths = equivalent_path_candidates(basename, root=dirname)
 
-        # check number of paths with the same case
-        if len(cased_path_candidates(basename, root=dirname)) > 1:
+        if len(equivalent_paths) > 1:
+
+            # We have different file names that would map to the same normalized path!
+
+            conflict_path = next(p for p in equivalent_paths if p != event.local_path)
+
+            # Check if we have a case conflict or a unicode conflict.
+
+            if normalize_case(event.local_path) == normalize_case(conflict_path):
+                suffix = "case conflict"
+            elif normalize_unicode(event.local_path) == normalize_case(conflict_path):
+                suffix = "unicode conflict"
+            else:
+                suffix = "normalization conflict"
 
             local_path_cc = generate_cc_name(
                 event.local_path,
-                suffix="case conflict",
-                is_fs_case_sensitive=self.is_case_sensitive,
+                suffix=suffix,
             )
 
             event_cls = DirMovedEvent if osp.isdir(event.local_path) else FileMovedEvent
@@ -2095,7 +2112,9 @@ class SyncEngine:
                 self.rescan(local_path_cc)
 
             self._logger.info(
-                'Case conflict: renamed "%s" to "%s"', event.local_path, local_path_cc
+                'Normalization conflict: renamed "%s" to "%s"',
+                event.local_path,
+                local_path_cc,
             )
 
             return True
@@ -2118,7 +2137,6 @@ class SyncEngine:
             local_path_cc = generate_cc_name(
                 event.local_path,
                 suffix="selective sync conflict",
-                is_fs_case_sensitive=self.is_case_sensitive,
             )
 
             event_cls = DirMovedEvent if osp.isdir(event.local_path) else FileMovedEvent
@@ -2238,7 +2256,7 @@ class SyncEngine:
 
         if self._handle_selective_sync_conflict(event):
             return None
-        if self._handle_case_conflict(event):
+        if self._handle_normalization_conflict(event):
             return None
 
         dbx_path_from = cast(str, event.dbx_path_from)
@@ -2258,9 +2276,9 @@ class SyncEngine:
 
         md_to_new = client.move(dbx_path_from, event.dbx_path, autorename=True)
 
-        self.remove_node_from_index(dbx_path_from)
+        self.remove_node_from_index(event.dbx_path_from_lower)
 
-        if md_to_new.path_lower != event.dbx_path.lower():
+        if md_to_new.name != osp.basename(event.local_path):
             # TODO: test this
             # conflicting copy created during upload, mirror remote changes locally
             local_path_cc = self.to_local_path(md_to_new.path_display, client)
@@ -2269,10 +2287,8 @@ class SyncEngine:
                 with convert_api_errors():
                     move(event.local_path, local_path_cc, raise_error=True)
 
-            # Delete entry of old path but don't update entry for new path here. This
-            # will force conflict resolution on download in case of intermittent
-            # changes.
-            self.remove_node_from_index(event.dbx_path)
+            # Delete entry of old path.
+            self.remove_node_from_index(event.dbx_path_lower)
             self._logger.info(
                 'Upload conflict: renamed "%s" to "%s"',
                 event.dbx_path,
@@ -2280,10 +2296,11 @@ class SyncEngine:
             )
 
         else:
-            self._update_index_recursive(md_to_new, client)
             self._logger.debug(
                 'Moved "%s" to "%s" on Dropbox', dbx_path_from, event.dbx_path
             )
+
+        self._update_index_recursive(md_to_new, client)
 
         return md_to_new
 
@@ -2316,7 +2333,7 @@ class SyncEngine:
 
         if self._handle_selective_sync_conflict(event):
             return None
-        if self._handle_case_conflict(event):
+        if self._handle_normalization_conflict(event):
             return None
 
         self._wait_for_creation(event.local_path)
@@ -2348,7 +2365,7 @@ class SyncEngine:
                     self.update_index_from_dbx_metadata(md_old, client)
                     return None
 
-            local_entry = self.get_index_entry(event.dbx_path)
+            local_entry = self.get_index_entry(event.dbx_path_lower)
 
             if not local_entry:
                 # file is new to us, let Dropbox rename it if something is in the way
@@ -2379,7 +2396,7 @@ class SyncEngine:
                 )
                 return None
 
-        if md_new.path_lower != event.dbx_path.lower():
+        if md_new.name != osp.basename(event.local_path):
             # conflicting copy created during upload, mirror remote changes locally
             local_path_cc = self.to_local_path(md_new.path_display, client)
             event_cls = DirMovedEvent if osp.isdir(event.local_path) else FileMovedEvent
@@ -2387,19 +2404,17 @@ class SyncEngine:
                 with convert_api_errors():
                     move(event.local_path, local_path_cc, raise_error=True)
 
-            # Delete entry of old path but don't update entry for new path here. This
-            # will force conflict resolution on download in case of intermittent
-            # changes.
-            self.remove_node_from_index(event.dbx_path)
+            # Delete entry of old path
+            self.remove_node_from_index(event.dbx_path_lower)
             self._logger.debug(
                 'Upload conflict: renamed "%s" to "%s"',
                 event.dbx_path,
-                md_new.path_lower,
+                md_new.path_display,
             )
         else:
-            # everything went well, update index
-            self.update_index_from_dbx_metadata(md_new, client)
             self._logger.debug('Created "%s" on Dropbox', event.dbx_path)
+
+        self.update_index_from_dbx_metadata(md_new, client)
 
         return md_new
 
@@ -2435,9 +2450,10 @@ class SyncEngine:
                     "the same as on Dropbox",
                     event.dbx_path,
                 )
+                self.update_index_from_dbx_metadata(md_old, client)
                 return None
 
-        local_entry = self.get_index_entry(event.dbx_path)
+        local_entry = self.get_index_entry(event.dbx_path_lower)
 
         if not local_entry:
             self._logger.debug(
@@ -2464,8 +2480,8 @@ class SyncEngine:
             )
             return None
 
-        if md_new.path_lower != event.dbx_path.lower():
-            # conflicting copy created during upload, mirror remote changes locally
+        if md_new.name != osp.basename(event.local_path):
+            # Conflicting copy created during upload, mirror remote changes locally.
             local_path_cc = self.to_local_path(md_new.path_display, client)
             with self.fs_events.ignore(FileMovedEvent(event.local_path, local_path_cc)):
                 try:
@@ -2474,18 +2490,18 @@ class SyncEngine:
                     with self.fs_events.ignore(FileDeletedEvent(event.local_path)):
                         delete(event.local_path)
 
-            # Delete revs of old path but don't set revs for new path here. This will
-            # force conflict resolution on download in case of intermittent changes.
-            self.remove_node_from_index(event.dbx_path)
+            # Delete revs of old path.
+            self.remove_node_from_index(event.dbx_path_lower)
             self._logger.debug(
                 'Upload conflict: renamed "%s" to "%s"',
                 event.dbx_path,
-                md_new.path_lower,
+                md_new.path_display,
             )
         else:
-            # everything went well, save new revs
-            self.update_index_from_dbx_metadata(md_new, client)
+            # everything went well
             self._logger.debug('Uploaded modified "%s" to Dropbox', md_new.path_lower)
+
+        self.update_index_from_dbx_metadata(md_new, client)
 
         return md_new
 
@@ -2508,13 +2524,13 @@ class SyncEngine:
         if event.local_path == self.dropbox_path:
             self.ensure_dropbox_folder_present()
 
-        if self.is_excluded_by_user(event.dbx_path):
+        if self.is_excluded_by_user(event.dbx_path_lower):
             self._logger.debug(
                 'Not deleting "%s": is excluded by selective sync', event.dbx_path
             )
             return None
 
-        local_rev = self.get_local_rev(event.dbx_path)
+        local_rev = self.get_local_rev(event.dbx_path_lower)
 
         md = client.get_metadata(event.dbx_path, include_deleted=True)
 
@@ -2525,14 +2541,16 @@ class SyncEngine:
                 md.path_display,
             )
             # don't delete a remote file if it was modified since last sync
-            if md.server_modified.timestamp() >= self.get_last_sync(event.dbx_path):
+            if md.server_modified.timestamp() >= self.get_last_sync(
+                event.dbx_path_lower
+            ):
                 self._logger.debug(
                     'Skipping deletion: remote item "%s" has been modified '
                     "since last sync",
                     md.path_display,
                 )
                 # mark local folder as untracked
-                self.remove_node_from_index(event.dbx_path)
+                self.remove_node_from_index(event.dbx_path_lower)
                 return None
 
         if event.is_file and isinstance(md, FolderMetadata):
@@ -2546,7 +2564,7 @@ class SyncEngine:
                 md.path_display,
             )
             # mark local file as untracked
-            self.remove_node_from_index(event.dbx_path)
+            self.remove_node_from_index(event.dbx_path_lower)
             return None
 
         try:
@@ -2568,7 +2586,7 @@ class SyncEngine:
             md_deleted = None
 
         # remove revision metadata
-        self.remove_node_from_index(event.dbx_path)
+        self.remove_node_from_index(event.dbx_path_lower)
 
         return md_deleted
 
@@ -2601,14 +2619,16 @@ class SyncEngine:
 
             md = client.get_metadata(dbx_path, include_deleted=True)
 
+            dbx_path_lower = normalize(dbx_path)
+
             if md is None:
                 # create a fake deleted event
-                index_entry = self.get_index_entry(dbx_path)
+                index_entry = self.get_index_entry(dbx_path_lower)
                 cased_path = index_entry.dbx_path_cased if index_entry else dbx_path
 
                 md = DeletedMetadata(
                     name=osp.basename(dbx_path),
-                    path_lower=dbx_path.lower(),
+                    path_lower=dbx_path_lower,
                     path_display=cased_path,
                 )
 
@@ -2841,7 +2861,7 @@ class SyncEngine:
                 new_excluded = [
                     path
                     for path in self.excluded_items
-                    if not is_equal_or_child(path, event.dbx_path.lower())
+                    if not is_equal_or_child(path, event.dbx_path_lower)
                 ]
 
                 self.excluded_items = new_excluded
@@ -3013,7 +3033,7 @@ class SyncEngine:
         items_to_discard = []
 
         for item in changes:
-            if self.is_excluded_by_user(item.dbx_path) or self.is_excluded(
+            if self.is_excluded_by_user(item.dbx_path_lower) or self.is_excluded(
                 item.dbx_path
             ):
                 items_to_discard.append(item)
@@ -3041,7 +3061,7 @@ class SyncEngine:
         :raises MaestralApiError: For any issues when syncing the item.
         """
 
-        local_rev = self.get_local_rev(event.dbx_path)
+        local_rev = self.get_local_rev(event.dbx_path_lower)
 
         if event.rev == local_rev:
             # Local change has the same rev. The local item (or deletion) must be newer
@@ -3061,7 +3081,7 @@ class SyncEngine:
             )
             return Conflict.Identical
         elif any(
-            is_equal_or_child(p, event.dbx_path.lower()) for p in self.upload_errors
+            is_equal_or_child(p, event.dbx_path_lower) for p in self.upload_errors
         ):
             # Local version could not be uploaded due to a sync error. Do not over-
             # write unsynced changes but declare a conflict.
@@ -3107,8 +3127,8 @@ class SyncEngine:
             # excluded names such as .DS_Store etc never count as unsynced changes
             return False
 
-        dbx_path = self.to_dbx_path(local_path)
-        index_entry = self.get_index_entry(dbx_path)
+        dbx_path_lower = self.to_dbx_path_lower(local_path)
+        index_entry = self.get_index_entry(dbx_path_lower)
 
         with convert_api_errors():  # catch OSErrors
 
@@ -3132,16 +3152,16 @@ class SyncEngine:
                             if self._ctime_newer_than_last_sync(entry.path):
                                 return True
                         elif not self.is_excluded(entry.name):
-                            child_dbx_path = self.to_dbx_path(entry.path)
+                            child_dbx_path_lower = self.to_dbx_path_lower(entry.path)
                             ctime = entry.stat().st_ctime
-                            if ctime > self.get_last_sync(child_dbx_path):
+                            if ctime > self.get_last_sync(child_dbx_path_lower):
                                 return True
 
                 return False
 
             else:
                 # Check our ctime against index.
-                return stat.st_ctime > self.get_last_sync(dbx_path)
+                return stat.st_ctime > self.get_last_sync(dbx_path_lower)
 
     def _get_ctime(self, local_path: str) -> float:
         """
@@ -3324,9 +3344,7 @@ class SyncEngine:
         # re-check for conflict and move the conflict
         # out of the way if anything has changed
         if self._check_download_conflict(event) == Conflict.Conflict:
-            new_local_path = generate_cc_name(
-                local_path, is_fs_case_sensitive=self.is_case_sensitive
-            )
+            new_local_path = generate_cc_name(local_path)
             event_cls = DirMovedEvent if osp.isdir(local_path) else FileMovedEvent
             with self.fs_events.ignore(event_cls(local_path, new_local_path)):
                 with convert_api_errors():
@@ -3342,7 +3360,7 @@ class SyncEngine:
                 delete(local_path)
 
         # check if we should preserve permissions of destination file
-        old_entry = self.get_index_entry(event.dbx_path)
+        old_entry = self.get_index_entry(event.dbx_path_lower)
 
         preserve_permissions = bool(old_entry and event.dbx_id == old_entry.dbx_id)
 
@@ -3400,9 +3418,7 @@ class SyncEngine:
             return None
 
         if conflict_check == Conflict.Conflict:
-            new_local_path = generate_cc_name(
-                event.local_path, is_fs_case_sensitive=self.is_case_sensitive
-            )
+            new_local_path = generate_cc_name(event.local_path)
             event_cls = DirMovedEvent if osp.isdir(event.local_path) else FileMovedEvent
             with self.fs_events.ignore(event_cls(event.local_path, new_local_path)):
                 with convert_api_errors():
@@ -3486,7 +3502,7 @@ class SyncEngine:
         """
 
         with self._database_access():
-            entry = self.get_index_entry(event.dbx_path.lower())
+            entry = self.get_index_entry(event.dbx_path_lower)
 
         if entry and entry.dbx_path_cased != event.dbx_path:
 
@@ -3521,8 +3537,6 @@ class SyncEngine:
 
             # add created and deleted events of children as appropriate
 
-            local_path_lower = local_path.lower()
-
             for path, stat in walk(local_path, self._scandir_with_ignore):
 
                 if S_ISDIR(stat.st_mode):
@@ -3531,6 +3545,8 @@ class SyncEngine:
                     self.fs_events.queue_event(FileModifiedEvent(path))
 
             # add deleted events
+
+            local_path_lower = normalize(local_path)
 
             with self._database_access():
 
@@ -3549,9 +3565,9 @@ class SyncEngine:
                         self.fs_events.queue_event(FileDeletedEvent(child_path))
 
         elif not osp.exists(local_path):
-            dbx_path = self.to_dbx_path(local_path)
+            dbx_path_lower = self.to_dbx_path_lower(local_path)
 
-            local_entry = self.get_index_entry(dbx_path)
+            local_entry = self.get_index_entry(dbx_path_lower)
 
             if local_entry:
                 if local_entry.is_directory: